backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.28.0
httpx>=0.24.0
lxml>=4.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0